+++ order/src/main/java/kitchenpos/presentation/TableGroupRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.presentation;
 
 import kitchenpos.application.OrderTableGroupService;
 import kitchenpos.domain.OrderTableGroup;
