+++ product/src/main/java/kitchenpos/product/domain/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.domain.repository;
+package kitchenpos.product.domain.repository;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
