+++ product/src/main/java/kitchenpos/product/dto/request/ProductCreateRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto.request;
+package kitchenpos.product.dto.request;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
