+++ product/src/main/java/kitchenpos/product/dto/response/ProductResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto.response;
+package kitchenpos.product.dto.response;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
