+++ settings.gradle
@@ -1,4 +1,5 @@
 rootProject.name = 'kitchenpos'
 include 'menu'
 include 'order'
+include 'product'
